+++ config/filament-trumbowyg.php
@@ -2,4 +2,14 @@
 
 return [
     'buttons' => [],
+    'tag_classes' => [],
+    'change_active_dropdown_icon' => false,
+    'semantic' => true,
+    'remove_format_pasted' => false,
+    'tags_to_remove' => [],
+    'tags_to_keep' => [],
+    'url_protocol' => false,
+    'minimal_links' => false,
+    'link_targets' => ['_self', '_blank'],
+    'image_width_modal_edit' => false,
 ];