+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to `filament-trumbowyg` will be documented in this file.
 
+## 2.1.0 - 2023-07-23
+- Adds support for tag classes [https://alex-d.github.io/Trumbowyg/documentation/#tag-classes]
+- Adds support for active dropdown icon [https://alex-d.github.io/Trumbowyg/documentation/#change-active-dropdown-icon]
+- Adds support for remove format on pasted [https://alex-d.github.io/Trumbowyg/documentation/#remove-format-pasted]
+- Adds support for removing tags [https://alex-d.github.io/Trumbowyg/documentation/#tags-to-remove]
+- Adds support for keeping tags [https://alex-d.github.io/Trumbowyg/documentation/#tags-to-keep]
+- Adds support for minimal links [https://alex-d.github.io/Trumbowyg/documentation/#minimal-links]
+- Adds support for link targets [https://alex-d.github.io/Trumbowyg/documentation/#link-targets]
+- Adds support for image width modal edit [https://alex-d.github.io/Trumbowyg/documentation/#image-width-modal-edit]
+- Adds support for url protocol [https://alex-d.github.io/Trumbowyg/documentation/#url-protocol]
+- Fix state issue when using a custom label
+
 ## 2.0.0 - 2023-04-12
 - Adds support for Laravel 10
 - Fix an issue accessing the editor data through `livewire.data`